apps/community/server/requirements.txt
# 本地开发引用 - 已移除Node.js包引用
# 注意：shared-types, shared-utils, shared-services 是Node.js包，不适用于Python项目

# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
httpx==0.25.2
aiohttp==3.9.1
python-dotenv==1.0.0
pydantic-settings==2.0.3

# Database
asyncpg==0.30.0
sqlalchemy[asyncio]>=2.0.25
alembic==1.12.1

# Redis cache
redis[hiredis]==5.0.1
aioredis==2.0.1

# Authentication & Security
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography>=41.0.0

# Validation & Serialization
pydantic[email]==2.10.0
email-validator==2.1.0

# Logging & Monitoring
structlog==23.2.0
python-json-logger==2.0.7

# Utilities
python-slugify==8.0.1
pytz==2023.3
pendulum==3.0.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-xdist==3.5.0
pytest-mock==3.12.0
pytest-benchmark==4.0.0
pytest-timeout==2.2.0
pytest-html==4.1.1
pytest-json-report==1.5.0
pytest-sugar==0.9.7
pytest-clarity==1.0.1
pytest-picked==0.5.0
factory-boy==3.3.0
faker==20.1.0
responses==0.24.1
# httpx-mock==0.11.0  # 暂时注释掉，版本不存在
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
bandit==1.7.5
# safety==2.3.5  # 暂时注释掉，与其他包存在packaging版本冲突

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# Production
gunicorn==21.2.0
prometheus-client==0.19.0